# Optimize default context propagation cases