--- a/src/Datadog.Trace/Propagation/ContextPropagatorBuilder.cs
+++ b/src/Datadog.Trace/Propagation/ContextPropagatorBuilder.cs
@@ -1,29 +1,24 @@
 // Modified by SignalFx
 
 using System;
-using System.Collections.Generic;
 using SignalFx.Tracing.Configuration;
 
 namespace SignalFx.Tracing.Propagation
 {
     internal static class ContextPropagatorBuilder
     {
-        private static readonly IReadOnlyDictionary<PropagatorType, Func<IPropagator>> _propagatorSelector =
-            new Dictionary<PropagatorType, Func<IPropagator>>()
-            {
-                { PropagatorType.B3, () => new B3SpanContextPropagator() },
-                { PropagatorType.Default, () => new B3SpanContextPropagator() },
-                { PropagatorType.W3C, () => W3CSpanContextPropagator.Instance }
-            };
-
         public static IPropagator BuildPropagator(PropagatorType propagator)
         {
-            if (_propagatorSelector.TryGetValue(propagator, out var getter))
+            switch (propagator)
             {
-                return getter();
+                case PropagatorType.B3:
+                case PropagatorType.Default:
+                    return new B3SpanContextPropagator();
+                case PropagatorType.W3C:
+                    return W3CSpanContextPropagator.Instance;
+                default:
+                    throw new InvalidOperationException($"There is no propagator registered for type '{propagator}'");
             }
-
-            throw new InvalidOperationException($"There is no propagator registered for type '{propagator}'");
         }
     }
 }