--- a/src/Datadog.Trace/Propagation/PropagationExtensions.cs
+++ b/src/Datadog.Trace/Propagation/PropagationExtensions.cs
@@ -17,24 +17,6 @@ namespace SignalFx.Tracing.Propagation
             return propagator.Extract(headers, ExtractFromHeadersCollection);
         }
 
-        public static IEnumerable<KeyValuePair<string, string>> ExtractHeaderTags(this IHeadersCollection headers, IEnumerable<KeyValuePair<string, string>> headerToTagMap)
-        {
-            foreach (KeyValuePair<string, string> headerNameToTagName in headerToTagMap)
-            {
-                var headerValue = ParseString(headers, headerNameToTagName.Key);
-
-                if (headerValue != null)
-                {
-                    yield return new KeyValuePair<string, string>(headerNameToTagName.Value, headerValue);
-                }
-            }
-        }
-
-        public static string ParseString(this IHeadersCollection headers, string headerName)
-        {
-            return PropagationHelpers.ParseString(headers, (carrier, header) => carrier.GetValues(header), headerName);
-        }
-
         private static void InjectToHeadersCollection(IHeadersCollection carrier, string header, string value)
         {
             carrier.Set(header, value);