Optimizing the default propagation using B3 by focusing on "common" cases:

1. Inbound request but no B3 headers: in this case bailout earlier avoiding allocations and logging on every attempt to extract - the big saving, in this case, are due to the removal of one log per "extract"
2. Inbound request with B3 headers: saving on allocations by removing "early" allocations and enumerator allocations (not a big save in speed in this case, but some memory allocation)

Original:

|   Method |      Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |----------:|----------:|----------:|-------:|------:|------:|----------:|
|     NoB3 | 18.678 us | 0.3681 us | 0.9825 us | 0.4578 |     - |     - |   1.88 KB |
| B3Header |  1.193 us | 0.0235 us | 0.0344 us | 0.2708 |     - |     - |   1.11 KB |

Optimized:

|   Method |       Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |-----------:|---------:|---------:|-------:|------:|------:|----------:|
|     NoB3 |   419.0 ns |  8.23 ns | 15.47 ns | 0.0801 |     - |     - |     336 B |
| B3Header | 1,003.6 ns | 19.72 ns | 35.55 ns | 0.1926 |     - |     - |     808 B |