@pjanotti 2021-06-10
> What have you used for benchmarking. https://github.com/dotnet/BenchmarkDotNet ?

Exactly.